% Auto-generated: do not edit by hand
\name{pagenotfoundThreejs}

\alias{pagenotfoundThreejs}

\title{PagenotfoundThreejs component}

\description{

}

\usage{
pagenotfoundThreejs(id=NULL)
}

\arguments{
\item{id}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

